Refactored to centralize shared code
