+++ src/AspNetCore.Swagger.Themes.Common/AspNetCore/Swagger/Themes/MimeTypes.cs
@@ -1,4 +1,4 @@
-﻿namespace AspNetCore.SwaggerUI.Themes;
+﻿namespace AspNetCore.Swagger.Themes;
 
 /// <summary>
 /// Standard MIME type constants.
