+++ src/AspNetCore.Swagger.Themes.Common/AspNetCore/Swagger/Themes/Scripts/modern.js
