+++ src/AspNetCore.Swagger.Themes.Common/AspNetCore/Swagger/Themes/Style.cs
@@ -1,4 +1,4 @@
-﻿namespace AspNetCore.NSwag.Themes;
+﻿namespace AspNetCore.Swagger.Themes;
 
 /// <summary>
 /// Represents a style for Swagger UI.
