+++ src/AspNetCore.Swagger.Themes.Common/AspNetCore/Swagger/Themes/Styles/desert.css
