+++ src/AspNetCore.Swagger.Themes.Common/AspNetCore/Swagger/Themes/Styles/modern.dark.css
