+++ src/AspNetCore.Swagger.Themes.Common/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("AspNetCore.SwaggerUI.Themes")]
+[assembly: InternalsVisibleTo("AspNetCore.NSwag.Themes")]
+[assembly: InternalsVisibleTo("AspNetCore.Swagger.Themes.Tests")]
