--- src/AspNetCore.SwaggerUI.Themes/AspNetCore/SwaggerUI/Themes/Styles/modern.light.css
@@ -1,7 +1,0 @@
-﻿/*
-    Modern Light Style
-
-    https://github.com/teociaps/SwaggerUI.Themes
-*/
-
-/* TODO: add custom style or continue to use common? */
