--- src/AspNetCore.SwaggerUI.Themes/Properties/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("AspNetCore.SwaggerUI.Themes.Tests")]
